\name{CropWatR}
\docType{package}
\title{
Crop-water balances according to algorithms described in the FAO Irrigation and drainage paper 56
}
\description{
Functions to enable calculation of FAO Penman-Monteith reference evapotranspiration, crop- and soil-water balances for agricultural crops. GIS-enabled implementation (using R's basic spatial packages, including rgeos, rgdal sp, raster) at a daily time-step, including calibration methods for irrigation scheduling and variable crop planting and harvesting dates. Crop-water models according to the FAO 56 document 'Crop evapotranspiration - Guidelines for computing crop water requirements'
}
\details{
\tabular{ll}{
Package: \tab CropWatR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-06\cr
License: \tab GPL-2\cr
}
}
\author{
Jacob Teter
Maintainer: <jacob.teter@gmail.com>
}
\references{
1. An improved crop-water model to estimate regional crop water use. Draft article for publication.
2. Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56 <http://www.fao.org/docrep/x0490e/x0490e00.htm>
}
\keyword{ spatial }
\seealso{
\code{\link[<pkg>:<sp>-package]{<https://cran.r-project.org/web/packages/sp/sp.pdf>}}
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
~~ simple examples of the most important functions ~~
}
